Update alpsdoocs_save.py
I removed the case check for a parent window within the child window as it was redundant, and the else case could have cause issues (it was from an earlier version where I called Tk.__init__() within SaveApp.__init__()"). The display text for the selected channels did not delete the previous selection before it printed the new ones. This was fixed. Finally, the class attribute "root" was renamed to "window" to be less ambiguous (since in some cases it could actually be the root window, but in other cases it might not be).