Update alpsdoocslib.py
removed instances of the print command in the get_data functions (including the progress bar), until they can be reimplemented to account for the GUI console box. a change to the continuous data processing function: append to a list the data from each channel from the current call to pydoocs.get_data, and send it all at once to the subroutine. this would allow more flexible data processing since all channels' data is available at once