Update alpsdoocs_analyze.py
now allow for fractional seconds as timebase as input, but the actual timebase may not be exactly as requested. we can only call pydoocs.read() an integer number of times. each call returns 500 data points, so we can only get an integer multiple of 500 data points. the "goal" for the number of returns of pydoocs.read() to get is rounded to the nearest multiple of 32 events per second (since the macropulses of DOOCS trigger 32 times every second)

instead of concatenating arrays, we append new arrays to a list and reshape at the end. this is was found to be 2 to 3 times faster